requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-dotenv>=1.0.1
python3-discogs-client>=2.7
anthropic>=0.49.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
youtube-search-python>=1.6.6
httpx>=0.25.0,<0.28.0  # youtube-search-python uses deprecated proxies kwarg removed in httpx 0.28

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.25.0,<0.28.0
